+++ src/api/dto/call.dto.ts
@@ -1,0 +1,8 @@
+export class Metadata {
+  number: string;
+}
+
+export class OfferCallDto extends Metadata {
+  isVideo?: boolean;
+  callDuration?: number;
+}
