+++ src/api/dto/sendMessage.dto.ts
@@ -46,11 +46,6 @@ export class Metadata {
   encoding?: boolean;
 }
 
-export class OfferCallDto extends Metadata {
-  isVideo?: boolean;
-  callDuration?: number;
-}
-
 export class SendTextDto extends Metadata {
   text: string;
 }
